LayoutTests/fast/dom/HTMLInputElement/input-size-attribute.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>

<body>
<script>
description('HTMLInputElement size attribute test');

var input = document.createElement("input");

shouldBe("input.size", "20");

shouldBe("input.setAttribute('size', '-1'); input.size", "20");
shouldBe("input.removeAttribute('size'); input.size", "20");
shouldBe("input.setAttribute('size', '1'); input.size", "1");
shouldBe("input.setAttribute('size', '2'); input.size", "2");
shouldBe("input.removeAttribute('size'); input.size", "20");
shouldBe("input.setAttribute('size', 'a'); input.size", "20");
shouldBe("input.removeAttribute('size'); input.size", "20");
shouldBe("input.setAttribute('size', '0'); input.size", "20");
shouldBe("input.setAttribute('size', '10'); input.size", "10");

shouldThrow("input.size = 0", "'Error: INDEX_SIZE_ERR: DOM Exception 1'");

</script>
</body>

<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
